AK+LibURL: Move CopyOnWrite<T> from LibURL to AK
(cherry picked from commit af68771dda8e28a27d93933d25f141e4752d35ad)
